+++ blocklets/ai-studio/src/components/publish/LogoField.tsx
@@ -1,8 +1,8 @@
 import { useUploader } from '@app/contexts/uploader';
-import PublishUpload from '@app/pages/project/icons/publish-upload';
+import Avatar from '@arcblock/ux/lib/Avatar';
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
 import { AssistantYjs } from '@blocklet/ai-runtime/types';
-import { Box, IconButton, Typography } from '@mui/material';
+import { Box, Stack, Typography } from '@mui/material';
 
 export default function LogoField({
   assistant,
@@ -17,44 +17,35 @@ export default function LogoField({
   return (
     <Box>
       <Typography mb={0.5} variant="subtitle2">
-        {t('publish.logo')}
+        {t('agentLogo')}
       </Typography>
 
-      <Box mb={0.5}>
+      <Stack
+        sx={{ cursor: 'pointer', gap: 0.5 }}
+        onClick={() => {
+          // @ts-ignore
+          const uploader = uploaderRef?.current?.getUploader();
+
+          uploader?.open();
+
+          uploader.onceUploadSuccess((data: any) => {
+            const { response } = data;
+            const url = response?.data?.url || response?.data?.fileUrl;
+            setRelease((release) => (release.logo = url));
+          });
+        }}>
         <Box
-          onClick={() => {
-            // @ts-ignore
-            const uploader = uploaderRef?.current?.getUploader();
+          component={Avatar}
+          src={assistant.release?.logo}
+          did={window.blocklet.appId}
+          size={100}
+          sx={{ borderRadius: 1 }}
+        />
 
-            uploader?.open();
-
-            uploader.onceUploadSuccess((data: any) => {
-              const { response } = data;
-              const url = response?.data?.url || response?.data?.fileUrl;
-              setRelease((release) => (release.logo = url));
-            });
-          }}>
-          {assistant.release?.logo ? (
-            <Box
-              component="img"
-              src={assistant.release.logo}
-              alt=""
-              sx={{ width: 100, height: 100, borderRadius: 1, cursor: 'pointer' }}
-            />
-          ) : (
-            <Box width={1} display="flex" gap={1.5}>
-              <IconButton
-                key="uploader-trigger"
-                size="small"
-                sx={{ borderRadius: 1, bgcolor: 'rgba(0, 0, 0, 0.06)', width: 100, height: 100 }}>
-                <PublishUpload sx={{ fontSize: 100 }} />
-              </IconButton>
-
-              <Typography variant="subtitle3">Supported file formats include PNG, JPG, and SVG.</Typography>
-            </Box>
-          )}
-        </Box>
-      </Box>
+        <Typography variant="caption" color="text.secondary">
+          {t('clickToUploadAgentLogo')}
+        </Typography>
+      </Stack>
     </Box>
   );
 }